feat: use readline module
